Revert "Refactor repository"
