+++ array_processing/algorithms/__init__.py
@@ -1,2 +1,1 @@
 from .array_algorithms import wlsqva, fk_freq
-from .helpers import wlsqva_proc, array_plot, getrij
