+++ array_processing/tools/arrayCharacteristics.py
@@ -1,12 +1,8 @@
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy import optimize
-from scipy.special import gammainc
 
-
-"""
-array_characterization
-----------------------
+# -*- coding: utf-8 -*-
+r"""
+arrayCharacteristics
+--------------------
 Provides
     1. Array characteristic methods applicable to geophysical sesnor arrays
     2. Support functions for array characteristic calculations
@@ -14,10 +10,10 @@ Provides
 How to use the module
 ---------------------
 Documentation is available in docstrings provided with the code. The
-docstring examples assume that `array_characterization` has been imported
+docstring examples assume that `arrayCharacteristics` has been imported
 as `arrChar`::
 
-    import array_characterization as arrChar
+    import arrayCharacteristics as arrChar
 
 Code snippets are indicated by three greater-than signs::
 
@@ -34,46 +30,56 @@ Each of the module's methods may be called as::
 
 or imported individually and called as::
 
-    >>> from array_characterization import co_array
+    >>> from arrayCharacteristics import co_array
     >>> dij = co_array(rij)
 
 Available methods
 -----------------
 arraySig
     Estimate array uncertainties and impulse response
-impulseResp
-    Calculate impulse response of an array
-rthEllipse
-    Calculate angles subtending and extremal distances to an ellipse       
-co_array
-    Form co-array coordinates from array coordinates
-chi2
-    Calculates value of :math:`\chi^2` for given confidence level    
-quadraticEqn
-    Find roots of :math:`ax^2 + bx + c = 0`    
-cubicEqn
-    Find roots of :math:`x^3 + ax^2 + bx + c = 0`
-quarticEqn
-    Find roots of :math:`x^4 + ax^3 + bx^2 + cx + d = 0`
 arraySigPlt
     Plots output of arraySig method
-arraySigContourPlt
-    Plots output of arraySig method onto a polar plot for a specified trace
-    velocity
+chi2
+    Calculates value of :math:`\chi^2` for given confidence level
+co_array
+    Form co-array coordinates from array coordinates
+cubicEqn
+    Find roots of :math:`x^3 + ax^2 + bx + c = 0`
+impulseResp
+    Calculate impulse response of an array
+quadraticEqn
+    Find roots of :math:`ax^2 + bx + c = 0`
+quarticEqn
+    Find roots of :math:`x^4 + ax^3 + bx^2 + cx + d = 0`
+rthEllipse
+    Calculate angles subtending and extremal distances to an ellipse
 
 Notes
 -----
 All of the methods in this module are written for use under Python 3.*
 
+"""
+
+# Archival note: per PEP 396, for a module with version-numbered methods,
+# the module has no version number.  Each method in this module has a
+# version-number embedded as an attribute per PEP 440 as (e.g.):
+#       method.__version__ = '3.2.1'
+
+
+"""
+Created on Wed Sep 18 13:02:15 2019
 @author: cas
       (c) 2019 Curt A. L. Szuberla
       University of Alaska Fairbanks, all rights reserved
 """
 
+# unicode test & digital signature of this module
+wolf = '\U0001F43A'
+print(wolf)
 
 def arraySig(rij, kmax, sigLevel, p=0.9, velLims=(0.27, 0.36), NgridV=100,
              NgridTh=100, NgridK=100):
-    """
+    r"""
     Estimate 2D array uncertainties in trace velocity and back azimuth,
     calculate impulse response
 
@@ -117,36 +123,47 @@ def arraySig(rij, kmax, sigLevel, p=0.9, velLims=(0.27, 0.36), NgridV=100,
         Vector of back azimuths :math:`(^\circ)` for axis in (NgridTh, ) array
     kvec : array
         Vector wavenumbers for axes in k-space in (NgridK, ) array
-    """
 
+
+    Version
+    -------
+    2.4 -- 20 Sep 2019
+
+    """
+    arraySig.__version__ = '2.4'
     # (c) 2019 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
+
+    from numpy import (linspace, pi, meshgrid, sin, cos, zeros, arctan2, array,
+                       sqrt, abs, diff)
+    from numpy.linalg import eig
+
 
     # calculate uncertainties
     # preliminaries
     dij = co_array(rij)
-    th = np.linspace(0, 360 * (1 - 1 / NgridTh), NgridTh) / 180 * np.pi
+    th = linspace(0, 360*(1-1/NgridTh), NgridTh) / 180*pi
     if len(velLims) == 1:
         vel = velLims
     else:
-        vel = np.linspace(velLims[0], velLims[1], NgridV)
-    Th, Vel = np.meshgrid(th, vel)
-    S1 = np.sin(Th) / Vel
-    S2 = np.cos(Th) / Vel
-    sigTh = np.zeros(Th.shape)
+        vel = linspace(velLims[0], velLims[1], NgridV)
+    Th, Vel = meshgrid(th, vel)
+    S1 = sin(Th) / Vel
+    S2 = cos(Th) / Vel
+    sigTh = zeros(Th.shape)
     sigV = sigTh.copy()
     # single-pass calcs
     # calculate eigenvalues/vectors of design matrix (one-time shot)
     C = dij@dij.T
-    cii, Ve = np.linalg.eig(C)
-    thEigR = np.arctan2(Ve[1, 0], Ve[0, 0])
-    R = np.array([[np.cos(thEigR), np.sin(thEigR)], [-np.sin(thEigR), np.cos(thEigR)]])
+    cii, Ve = eig(C)
+    thEigR = arctan2(Ve[1, 0], Ve[0, 0])
+    R = array([[cos(thEigR), sin(thEigR)], [-sin(thEigR), cos(thEigR)]])
     # calculate chi2 for desired confidence level
     x2 = chi2(2, 1-p)
-    sigS = sigLevel / np.sqrt(cii)
+    sigS = sigLevel / sqrt(cii)
     # prep for loop
-    a = np.sqrt(x2) * sigS[0]
-    b = np.sqrt(x2) * sigS[1]
+    a = sqrt(x2)*sigS[0]
+    b = sqrt(x2)*sigS[1]
     N, M = Th.shape
     # froot loops
     for n in range(N):
@@ -157,15 +174,15 @@ def arraySig(rij, kmax, sigLevel, p=0.9, velLims=(0.27, 0.36), NgridV=100,
             # rotate & recalculate
             eVec = eVec @ R
             # fix up angle calcs
-            sigTh[n, m] = np.abs(np.diff(
-                (np.arctan2(eVec[2:, 1], eVec[2:, 0]) * 180 / np.pi - 360) % 360
+            sigTh[n, m] = abs(diff(
+                    (arctan2(eVec[2:, 1], eVec[2:, 0])*180/pi-360)%360
                     ))
             if sigTh[n, m] > 180:
-                sigTh[n, m] = np.abs(sigTh[n, m] - 360)
-            sigV[n, m] = np.abs(np.diff(1 / eExtrm[:2]))
+                sigTh[n, m] = abs(sigTh[n, m] - 360)
+            sigV[n, m] = abs(diff(1 / eExtrm[:2]))
     # prepare impulse response
     impResp, kvec = impulseResp(dij, kmax, NgridK)
-    return sigV, sigTh, impResp, vel, th / np.pi * 180, kvec
+    return sigV, sigTh, impResp, vel, th/pi*180, kvec
 
 
 def impulseResp(dij, kmax, NgridK):
@@ -188,21 +205,27 @@ def impulseResp(dij, kmax, NgridK):
         Impulse response over grid as (NgridK, NgridK) array
     kvec : array
         Vector wavenumbers for axes in k-space in (NgridK, ) array
-    """
 
+    Version
+    ~~~~~~~
+    1.0 -- 20 Sep 2019
+
+    """
+    impulseResp.__version__ = '1.0'
     # (c) 2019 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
 
+    from numpy import linspace, meshgrid, vstack, cos, sum, reshape, sqrt
 
 
     # pre-allocate grid for k-space
-    kvec  = np.linspace(-kmax, kmax, NgridK)
-    Kx, Ky = np.meshgrid(kvec, kvec)
+    kvec  = linspace(-kmax, kmax, NgridK)
+    Kx, Ky = meshgrid(kvec, kvec)
     N = dij.shape[1]
-    K = np.vstack((Ky.flatten(), Kx.flatten())).T
-    d = 2 * np.cos(K @ dij)
+    K = vstack((Ky.flatten(), Kx.flatten())).T
+    d = 2 * cos(K@dij)
     # last term adds in fact that cos(0)==1 for ignored self-delay terms
-    d = np.reshape(np.sum(d, axis=1), (NgridK, NgridK)) + (1 + np.sqrt(1 + 8 * N)) / 2
+    d = reshape(sum(d, axis=1), (NgridK, NgridK)) + (1+sqrt(1 + 8*N))/2
     return d, kvec
 
 
@@ -232,12 +255,18 @@ def rthEllipse(a, b, x0, y0):
         [[x min dist., y min dist.], [x max dist., y max dist.],
          [x max angle tangency, y max angle tangency],
          [x min angle tangency, y min angle tangency]]
-    """
 
+    Version
+    ~~~~~~~
+    1.1.1 -- 7 Oct 2019
+
+    """
+    rthEllipse.__version__ = '1.1.1'
     # (c) 2019 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
 
-
+    from numpy import (zeros, abs, finfo, float64, array, sqrt, where, real,
+                       vstack, max, hstack, sort, arctan2, pi)
 
     # set constants
     A = 2/a**2
@@ -248,9 +277,9 @@ def rthEllipse(a, b, x0, y0):
     F = C-A
     G = A/2
     H = C/2
-    eExtrm = np.zeros((4,))
-    eVec = np.zeros((4, 2))
-    eps = np.finfo(np.float64).eps
+    eExtrm = zeros((4, ))
+    eVec = zeros((4, 2))
+    eps = finfo(float64).eps
 
     # some tolerances for numerical errors
     circTol = 1e8  # is it circular to better than circTol*eps?
@@ -258,67 +287,67 @@ def rthEllipse(a, b, x0, y0):
     magTol = 1e-5  # is a sol'n within ellipse*(1+magTol) (maginification)
 
     # pursue circular or elliptical solutions
-    if np.abs(F) <= circTol * eps:
+    if abs(F) <= circTol * eps:
         # circle
-        cent = np.sqrt(x0 ** 2 + y0 ** 2)
-        eExtrm[0:2]  = cent + np.array([-a, a])
-        eVec[0:2, :] = np.array([
+        cent = sqrt(x0**2 + y0**2)
+        eExtrm[0:2]  = cent + array([-a, a])
+        eVec[0:2, :] = array([
                 [x0-a*x0/cent, y0-a*y0/cent],
                 [x0+a*x0/cent, y0+a*y0/cent]
                 ])
     else:
         # ellipse
         # check for trivial distance sol'n
-        if np.abs(y0) <  zeroTol * eps:
-            eExtrm[0:2] = x0 + np.array([-a, a])
-            eVec[0:2, :] = np.vstack((eExtrm[0:2], [0, 0])).T
-        elif np.abs(x0) <  zeroTol * eps:
-            eExtrm[0:2] = y0 + np.array([-b, b])
-            eVec[0:2, :] = np.vstack(([0, 0], eExtrm[0:2])).T
+        if abs(y0) <  zeroTol * eps:
+            eExtrm[0:2] = x0 + array([-a, a])
+            eVec[0:2, :] = vstack( (eExtrm[0:2],  [0, 0]) ).T
+        elif abs(x0) <  zeroTol * eps:
+            eExtrm[0:2] = y0 + array([-b, b])
+            eVec[0:2, :] = vstack( ([0, 0], eExtrm[0:2]) ).T
         else:
             # use dual solutions of quartics to find best, real-valued results
             # solve quartic for y
             fy = F**2*H
             y = quarticEqn(-D*F*(2*H+F)/fy, (B**2*(G+F)+E*F**2+D**2*(H+2*F))/fy,
                            -D*(B**2+2*E*F+D**2)/fy, (D**2*E)/fy)
-            y = np.array([y[i] for i in list(np.where(y == np.real(y))[0])])
+            y = array([y[i] for i in list(where(y == real(y))[0])])
             xy = B*y / (D-F*y)
             # solve quartic for x
             fx = F**2*G
             x = quarticEqn(B*F*(2*G-F)/fx, (B**2*(G-2*F)+E*F**2+D**2*(H-F))/fx,
                            B*(2*E*F-B**2-D**2)/fx, (B**2*E)/fx)
-            x = np.array([x[i] for i in list(np.where(x == np.real(x))[0])])
+            x = array([x[i] for i in list(where(x == real(x))[0])])
             yx = D*x / (F*x+B)
             # combine both approaches
-            distE = np.hstack(
-                    (np.sqrt(x ** 2 + yx ** 2), np.sqrt(xy ** 2 + y ** 2))
+            distE = hstack(
+                    (sqrt(x**2 + yx**2), sqrt(xy**2 + y**2))
                     )
             # trap real, but bogus sol's (esp. near Th = 180)
-            distEidx = np.where(
-                (distE <= np.sqrt(x0 ** 2 + y0 ** 2) + np.max([a, b]) * (1 + magTol)) &
-                (distE >= np.sqrt(x0 ** 2 + y0 ** 2) - np.max([a, b]) * (1 + magTol))
+            distEidx = where(
+                (distE <= sqrt(x0**2 + y0**2)+max([a, b])*(1+magTol)) &
+                (distE >= sqrt(x0**2 + y0**2)-max([a, b])*(1+magTol))
                 )
-            coords = np.hstack(((x, yx), (xy, y))).T
+            coords = hstack(((x, yx), (xy, y))).T
             coords = coords[distEidx,:][0]
             distE = distE[distEidx]
             eExtrm[0:2] = [distE.min(),  distE.max()]
-            eVec[0:2, :] = np.vstack(
-                    (coords[np.where(distE == distE.min()), :][0][0],
-                     coords[np.where(distE == distE.max()), :][0][0])
+            eVec[0:2, :] = vstack(
+                    (coords[where(distE == distE.min()), :][0][0],
+                     coords[where(distE == distE.max()), :][0][0])
                     )
     # angles subtended
     if x0 < 0:
         x0 = -x0
-        y = -np.array(quadraticEqn(D ** 2 + B ** 2 * H / G, 4 * D * E, 4 * E ** 2 - B ** 2 * E / G))
-        x = -np.sqrt(E / G - H / G * y ** 2)
+        y = -array(quadraticEqn(D**2 + B**2*H/G, 4*D*E, 4*E**2 - B**2*E/G))
+        x = -sqrt(E/G - H/G*y**2)
     else:
-        y = -np.array(quadraticEqn(D ** 2 + B ** 2 * H / G, 4 * D * E, 4 * E ** 2 - B ** 2 * E / G))
-        x = np.sqrt(E / G - H / G * y ** 2)
-    eVec[2:, :] = np.vstack((x, y)).T
+        y = -array(quadraticEqn(D**2 + B**2*H/G, 4*D*E, 4*E**2 - B**2*E/G))
+        x = sqrt(E/G - H/G*y**2)
+    eVec[2:, :] = vstack((x, y)).T
     # various quadrant fixes
-    if x0 == 0 or np.abs(x0) - a < 0:
+    if x0 == 0 or abs(x0) - a < 0:
         eVec[2, 0] = -eVec[2, 0]
-    eExtrm[2:] = np.sort(np.arctan2(eVec[2:, 1], eVec[2:, 0]) / np.pi * 180)
+    eExtrm[2:] = sort(arctan2(eVec[2:, 1],eVec[2:, 0])/pi*180)
     return eExtrm, eVec
 
 
@@ -336,8 +365,13 @@ def co_array(rij):
     ~~~~~~~
     dij : array
         (d, n(n-1)//2) co-array, coordinates of the sensor pairing separations
-    """
 
+    Version
+    ~~~~~~~
+    1.0 -- 13 Feb 2017
+
+    """
+    co_array.__version__ = '1.0'
     # (c) 2017 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
     #
@@ -347,7 +381,7 @@ def co_array(rij):
 
 
 def chi2(nu, alpha, funcTol=1e-10):
-    """
+    r"""
     Calculate value of a :math:`\chi^2` such that a :math:`\nu`-dimensional
     confidence ellipsoid encloses a fraction :math:`1 - \alpha` of normally
     distributed variable
@@ -367,25 +401,33 @@ def chi2(nu, alpha, funcTol=1e-10):
     chi2val : float
         value of a :math:`\chi^2` enclosing :math:`1 - \alpha` confidence
         region
+
+    Version
+    ~~~~~~~
+    1.0 -- 7 Oct 2019
+
     """
 
+    chi2.__version__ = '1.0'
     # (c) 2019 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
 
     if nu == 2:
-
+        from numpy import log
         # this shorthand owing to Ken Arnoult
-        return -2 * np.log(alpha)
+        return -2 * log(alpha)
     else:
         # but just in case we end up with a nu != 2 situation
-
-        gammaTest = lambda X2test: np.abs(gammainc(nu / 2,
-                                                      X2test / 2) - (1-alpha))
+        from scipy import optimize
+        from scipy.special import gammainc
+        from numpy import abs
+        gammaTest = lambda X2test: abs(gammainc(nu/2,
+                                                X2test/2) - (1-alpha))
         return optimize.fmin(func=gammaTest, x0=1, ftol=funcTol, disp=False)
 
 
 def quadraticEqn(a,b,c):
-    """
+    r"""
     Roots of quadratic equation in the form :math:`ax^2 + bx + c = 0`
 
     Parameters
@@ -406,32 +448,39 @@ def quadraticEqn(a,b,c):
     ~~~~~
     1) Stable solutions, even for :math:`b^2 >> ac` or complex coefficients,
     per algorithm of NR 2d ed. :math:`\S` 5.6.
-    """
 
+    Version
+    ~~~~~~~
+    1.0.1 -- 10 Feb 2017
+
+    """
+    quadraticEqn.__version__ = '1.0.1'
     # (c) 2017 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
     #
+    from numpy import isreal, sign, real, sqrt, complex
     # real coefficient branch
-    if np.isreal([a, b, c]).all():
+    if isreal([a, b, c]).all():
         # note np.sqrt(-1) = nan, so force complex argument
         if b:
             # std. sub-branch
-            q = -0.5*(b + np.sign(b) * np.sqrt(np.complex(b * b - 4 * a * c)))
+            q = -0.5*(b + sign(b)*sqrt(complex(b*b - 4*a*c)))
         else:
             # b = 0 sub-branch
-            q = -np.sqrt(np.complex(-a * c))
+            q = -sqrt(complex(-a*c))
     # complex coefficient branch
     else:
-        if np.real(np.conj(b) * np.sqrt(b * b - 4 * a * c)) >= 0:
-            q = -0.5*(b + np.sqrt(b * b - 4 * a * c))
+        from numpy import conj
+        if real(conj(b)*sqrt(b*b - 4*a*c)) >= 0:
+            q = -0.5*(b + sqrt(b*b - 4*a*c))
         else:
-            q = -0.5*(b - np.sqrt(b * b - 4 * a * c))
+            q = -0.5*(b - sqrt(b*b - 4*a*c))
     # stable root solution
     x = [q/a, c/q]
     # parse real and/or int roots for tidy output
     for k in 0,1:
-        if np.real(x[k]) == x[k]:
-            x[k] = float(np.real(x[k]))
+        if real(x[k]) == x[k]:
+            x[k] = float(real(x[k]))
             if int(x[k]) == x[k]:
                 x[k] = int(x[k])
     return x
@@ -460,50 +509,58 @@ def cubicEqn(a,b,c):
     1) Relatively stable solutions, with some tweaks by Dr. Z, per algorithm
     of NR 2d ed. :math:`\S` 5.6.  Even np.roots can have some (minor) issues;
     e.g., :math:`x^3 - 5x^2 + 8x - 4 = 0`.
-    """
 
+    Version
+    ~~~~~~~
+    1.0.1 -- 11 Feb 2017
+
+    """
     # (c) 2017 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
     #
-
+    cubicEqn.__version__ = '1.0.1'
+    from numpy import isreal, sqrt, real
     Q = a*a/9 - b/3
     R = (3*c - a*b)/6 + a*a*a/27
     Q3 = Q*Q*Q
     R2 = R*R
     ao3 = a/3
     # Q & R are real
-    if np.isreal([a, b, c]).all():
+    if isreal([a, b, c]).all():
         # 3 real roots
         if R2 < Q3:
-            sqQ = -2 * np.sqrt(Q)
-            theta = np.arccos(R / np.sqrt(Q3))
+            from numpy import arccos, cos, pi
+            sqQ = -2*sqrt(Q)
+            theta = arccos(R/sqrt(Q3))
             # this solution first published in 1615 by Viète!
             x = [
-                sqQ * np.cos(theta / 3) - ao3,
-                sqQ * np.cos((theta + 2 * np.pi) / 3) - ao3,
-                sqQ * np.cos((theta - 2 * np.pi) / 3) - ao3
+                    sqQ*cos(theta/3) - ao3,
+                    sqQ*cos((theta + 2*pi)/3) - ao3,
+                    sqQ*cos((theta - 2*pi)/3) - ao3
                     ]
         # Q & R real, but 1 real, 2 complex roots
         else:
             # this is req'd since np.sign(0) = 0
             if R != 0:
-                A = -np.sign(R) * (np.abs(R) + np.sqrt(R2 - Q3)) ** (1 / 3)
+                from numpy import sign
+                A = -sign(R)*(abs(R) + sqrt(R2 - Q3))**(1/3)
             else:
-                A = -np.sqrt(-Q3) ** (1 / 3)
+                A = -sqrt(-Q3)**(1/3)
             if A==0:
                 B = 0
             else:
                 B = Q/A
             # one real root & two conjugate complex ones
             x = [
-                (A+B) - ao3,
-                -.5 * (A+B) + 1j * np.sqrt(3) / 2 * (A - B) - ao3,
-                -.5 * (A+B) - 1j * np.sqrt(3) / 2 * (A - B) - ao3
+                    (A+B) - ao3,
+                    -.5*(A+B) + 1j*sqrt(3)/2*(A-B) - ao3,
+                    -.5*(A+B) - 1j*sqrt(3)/2*(A-B) - ao3
                     ]
     # Q & R complex, so also 1 real, 2 complex roots
     else:
-        sqR2mQ3 = np.sqrt(R2 - Q3)
-        if np.real(np.conj(R) * sqR2mQ3) >= 0:
+        sqR2mQ3 = sqrt(R2-Q3)
+        from numpy import conj
+        if real(conj(R)*sqR2mQ3) >= 0:
             A = -(R+sqR2mQ3)**(1/3)
         else:
             A = -(R-sqR2mQ3)**(1/3)
@@ -513,14 +570,14 @@ def cubicEqn(a,b,c):
             B = Q/A
         # one real root & two conjugate complex ones
         x = [
-            (A+B) - ao3,
-            -.5 * (A+B) + 1j * np.sqrt(3) / 2 * (A - B) - ao3,
-            -.5 * (A+B) - 1j * np.sqrt(3) / 2 * (A - B) - ao3
+                (A+B) - ao3,
+                -.5*(A+B) + 1j*sqrt(3)/2*(A-B) - ao3,
+                -.5*(A+B) - 1j*sqrt(3)/2*(A-B) - ao3
                 ]
     # parse real and/or int roots for tidy output
     for k in range(0,3):
-        if np.real(x[k]) == x[k]:
-            x[k] = float(np.real(x[k]))
+        if real(x[k]) == x[k]:
+            x[k] = float(real(x[k]))
             if int(x[k]) == x[k]:
                 x[k] = int(x[k])
     return x
@@ -549,26 +606,31 @@ def quarticEqn(a,b,c,d):
     ~~~~~
     1) Stable solutions per algorithm of CRC  Std. Mathematical
     Tables, 29th ed.
-    """
 
+    Version
+    ~~~~~~~
+    1.0.1 -- 11 Feb 2017
+
+    """
     # (c) 2017 Curt A. L. Szuberla
     # University of Alaska Fairbanks, all rights reserved
     #
-
+    quarticEqn.__version__ = '1.0.1'
+    from numpy import sqrt, real
     # find *any* root of resolvent cubic
     a2 = a*a
     y = cubicEqn(-b, a*c - 4*d, (4*b - a2)*d - c*c)
     y = y[0]
     # find R
-    R = np.sqrt(a2 / 4 - (1 + 0j) * b + y) # force complex in sqrt
+    R = sqrt(a2/4 - (1+0j)*b + y) # force complex in sqrt
     foo = 3*a2/4 - R*R - 2*b
     if R != 0:
-        D = np.sqrt(foo + (a * b - 2 * c - a2 * a / 4) / R) # R is already complex here
-        E = np.sqrt(foo - (a * b - 2 * c - a2 * a / 4) / R) # ...
+        D = sqrt(foo + (a*b - 2*c - a2*a/4)/R) # R is already complex here
+        E = sqrt(foo - (a*b - 2*c - a2*a/4)/R) # ...
     else:
-        sqrtTerm = 2 * np.sqrt(y * y - (4 + 0j) * d) # force complex in sqrt
-        D = np.sqrt(foo + sqrtTerm)
-        E = np.sqrt(foo - sqrtTerm)
+        sqrtTerm = 2*sqrt(y*y - (4+0j)*d) # force complex in sqrt
+        D = sqrt(foo + sqrtTerm)
+        E = sqrt(foo - sqrtTerm)
     x = [
             -a/4 + R/2 + D/2,
             -a/4 + R/2 - D/2,
@@ -577,8 +639,8 @@ def quarticEqn(a,b,c,d):
             ]
     # parse real and/or int roots for tidy output
     for k in range(0,4):
-        if np.real(x[k]) == x[k]:
-            x[k] = float(np.real(x[k]))
+        if real(x[k]) == x[k]:
+            x[k] = float(real(x[k]))
             if int(x[k]) == x[k]:
                 x[k] = int(x[k])
     return x
@@ -610,7 +672,14 @@ def arraySigPlt(rij, sig, sigV, sigTh, impResp, vel, th, kvec, figName=None):
         Vector wavenumbers for axes in k-space in (NgridK, ) array
     figName : str
         Name of output file, will be written as figName.png (optional)
+
     """
+
+    # this is just a sanity check, so no real docstring
+    arraySigPlt.__version__ = '1.0'
+
+    # std. (if inefficient) plotting import block
+    import matplotlib.pyplot as plt
 
     # for plotting methods & scripts
     figFormat = 'png'       # MUCH faster than pdf!!
@@ -682,6 +751,8 @@ def arraySigContourPlt(sigV, sigTh, vel, th, trace_v):
     author: D. Fee
 
     """
+    import matplotlib.pyplot as plt
+    import numpy as np
 
     tvel_ptr = np.abs(vel - trace_v).argmin()
     sigV_cont = sigV[tvel_ptr,:]
@@ -710,3 +781,4 @@ def arraySigContourPlt(sigV, sigTh, vel, th, trace_v):
     ax2.set_title('Back-Azimuth Uncertainty, V=%.2f' % trace_v, va='bottom', pad=20)
 
     return fig
+
